added decorate method and refactor
